# Make SchemaConfigured up to 100% faster