--- a/tox.ini
+++ b/tox.ini
@@ -15,6 +15,7 @@ basepython =
     python3
 commands =
     coverage run -m zope.testrunner --test-path=src []
+    coverage html
     coverage report --fail-under=100
 deps =
     {[testenv]deps}