Here are the results of testing the three commits in this PR using objects implementing 20 interfaces for a total of 20 schema fields. Benchmarks with `SC` in the name are SchemaConfigured, others are not and serve as a control. The `WideInheritance` class doesn't use any `FieldProperty` attributes, the others do.

| Benchmark                   | Master                 | Faster schemadict             | Cache schemadict              | Special-case FieldProperty      |
|-----------------------------|------------------------|-------------------------------|-------------------------------|---------------------------------|
| Create WideInheritance      | 63.8 ns                | 58.0 ns: 1.10x faster (-9%)   | 61.3 ns: 1.04x faster (-4%)   | 66.0 ns: 1.03x slower (+3%)     |
| Create SCWideInheritance    | 294 us                 | 39.8 us: 7.39x faster (-86%)  | 16.2 us: 18.11x faster (-94%) | 19.1 us: 15.42x faster (-94%)   |
| Create DeepestInheritance   | 77.8 ns                | 93.3 ns: 1.20x slower (+20%)  | 96.4 ns: 1.24x slower (+24%)  | 102 ns: 1.31x slower (+31%)     |
| Create SCDeepestInheritance | 654 us                 | 57.0 us: 11.47x faster (-91%) | 16.2 us: 40.23x faster (-98%) | 2.54 us: 257.52x faster (-100%) |
| Create ShallowInheritance   | 66.0 ns                | 57.9 ns: 1.14x faster (-12%)  | 61.1 ns: 1.08x faster (-7%)   | 63.7 ns: 1.04x faster (-4%)     |
| Create SCShallowInheritance | 304 us                 | 57.1 us: 5.33x faster (-81%)  | 33.1 us: 9.20x faster (-89%)  | 2.30 us: 132.53x faster (-99%)  |

Everything except for the very last commit ought to be totally safe (possibly modulo requiring the return from `schemadict` to be immutable). I *think* the last commit is safe as well.