@cutz 2020-05-02
In the benchmark numbers above are the timing changes for the _non SC_ classes just due to timing variances in the benchmark runs?