@jamadden 2020-05-02
And here's PyPy totally rocking it, getting another ~100% improvement.

| Benchmark                   | CPython 2.7 | PyPy 2 7.3.1                   |
|-----------------------------|---------------------------|-------------------------------|
| Create WideInheritance      | 59.4 ns                   | 1.31 ns: 45.42x faster (-98%) |
| Create SCWideInheritance    | 17.9 us                   | 2.53 us: 7.08x faster (-86%)  |
| Create DeepestInheritance   | 91.7 ns                   | 1.30 ns: 70.50x faster (-99%) |
| Create SCDeepestInheritance | 2.78 us                   | 512 ns: 5.44x faster (-82%)   |
| Create ShallowInheritance   | 57.3 ns                   | 1.32 ns: 43.30x faster (-98%) |
| Create SCShallowInheritance | 2.65 us                   | 509 ns: 5.21x faster (-81%)   |