@cutz 2020-05-04
> And here's PyPy totally rocking it

If only it were that simple for the system as a whole...