@jamadden 2020-05-04
(I’m working on a release of https://github.com/zopefoundation/z3c.recipe.compattest which should make it much easier to isolate problematic packages in a whole system. git master should be usable.)