@jamadden 2020-05-06
Thanks. I'm throwing up on PyPI now.